requirements.txt
# Requirements for NCBI Toolkit

# BioPython - Required for NCBI Entrez access and sequence handling
biopython>=1.79
